feat(catalyst-ui): inline edit-in-place on the catalog detail page (#3648, founder item #1)
"Editing the catalog item on a separate edit chip and popup is a very bad
idea — we already have the individual page and it must be editable."

Extracts the edit form into a reusable CatalogEditForm (no modal chrome);
CatalogDetail (/catalog/bp-<name>) renders it inline behind an admin "Edit"
toggle in the hero, saving via the same path + refreshing the page. Topology
checkboxes pre-check from the Blueprint's declared topologies (canonical →
editor-dialect map). CatalogEditDialog is reduced to a thin modal wrapper
around the shared form so existing tests stay green (typecheck clean, 24
catalog tests pass).

Follow-up (same PR): rewire the catalog-grid chip to navigate here + drop the
modal; and per founder #8 (IaC-first) converge the save from the commerce
overlay to a catalog-git commit.

Refs #3648.
